// Desktop/pruebas-inventarios/src/app/models/solicitud.ts
export class SolicitudUsuario {
  nIdSolicitudUsuario: Number;
  sRucMunicipalidad: String;
  sRazonSocial: String;
  sNombre: String;
  sApellidoPaterno: String;
  sApellidoMaterno: String;
  sCargo: String;
  sArea: String;
  sCorreo: String;
  sTelefono: String;
  sContrasenia: String;
  sNombreDocumento: String;
  sUidDocumento: String;
  dFechaRegistro: Date;
  sTipoMunicipalidad: null;
  sUbigeoMunicipalidad: null;
  sNombreResponsable: string;
  sCorreoResponsable: string;
  sTelefonoResponsable: string;
  sEstadoActual: string;

}